Update to Vert.x 4.0.0 EDGCOMMON-30
The OkapiClient in this module (one of dozens) has changed its API.
OkapiClient.{post,put,get,delete} has a response handler now of type
Handler<HttpResponse<Buffer>> it used to be: Handler<HttpClientResponse>.

The OkapiClient really doesn't give much over WebClient. And
we should consider just dropping it.

+++ src/test/java/org/folio/edge/core/utils/OkapiClientFactoryTest.java
@@ -10,7 +10,7 @@ import io.vertx.core.Vertx;
 
 public class OkapiClientFactoryTest {
 
-  private static final long reqTimeout = 5000L;
+  private static final int reqTimeout = 5000;
   private static final String tenant = "diku";
 
   private OkapiClientFactory ocf;
